video id change int to string. cache sorted with desc. gone for channel name in player.

+++ ets2mm/src/main/java/com/kyawhut/atsycast/ets2mm/ui/cache/CacheFragment.kt
@@ -70,7 +70,7 @@ internal class CacheFragment : BaseGridSupportFragment<CacheViewModel>() {
             )
         } else if (it is RecentlyWatchEntity) {
             startActivity<PlayerActivity>(
-                Constants.EXTRA_VIDEO_ID to it.videoID.toInt(),
+                Constants.EXTRA_VIDEO_ID to it.videoID,
                 Constants.EXTRA_IS_RESUME to true,
                 Constants.EXTRA_IS_ADULT to it.isAdult,
                 Constants.EXTRA_VIDEO_TITLE to it.videoTitle,
